rust: More properly/safely hide util::os from wasm
Rather than letting wasm-configured code to call
functions like home_dir_path, these are not simply
invislble in this mode, which will cause the build
to fail.

Moved around some dependent code as well, and made
the root cmd and command-line functions similarly
invisible to this build mode, since they're irrelevant.
