+++ src/app/outfmt.rs
@@ -1,3 +1,5 @@
-pub mod csv;
 pub mod model;
 pub mod text;
+
+#[cfg(not(target_arch = "wasm32"))]
+pub mod csv;
