+++ src/app/outfmt/csv.rs
@@ -1,7 +1,7 @@
 use std::{fs::File, io, path::PathBuf};
 
 
-use crate::util::sys::mk_writable_dir;
+use crate::util::os::mk_writable_dir;
 
 use super::model::{AcbWriter, OutputType};
