+++ src/lib.rs
@@ -1,10 +1,12 @@
 pub mod app;
-pub mod cmd;
 pub mod fx;
 pub mod log;
 pub mod portfolio;
 pub mod tracing;
 pub mod util;
+
+#[cfg(not(target_arch = "wasm32"))]
+pub mod cmd;
 
 extern crate lazy_static;
