+++ src/util.rs
@@ -1,6 +1,10 @@
+pub mod basic;
 pub mod date;
 pub mod decimal;
 pub mod math;
 pub mod rc;
 pub mod rw;
 pub mod sys;
+
+#[cfg(not(target_arch = "wasm32"))]
+pub mod os;
