+++ src/util/basic.rs
@@ -1,0 +1,1 @@
+pub type SError = String;
